THE BULLETS FINALLY ARE LIKE THEY WERE 8 MONTHS AGO!!

+++ Games/SpaceSand/include/Components/Projectile.h
@@ -4,7 +4,7 @@
 
 struct Projectile {
 	std::function<void()> FixedUpdate; // called in physics tick
-	std::function<std::tuple<bool, int, int>()> Update;      // called in render tick, return true if collision
+	std::function<std::tuple<bool, float, float>()> Update; // called in render tick, return true if collision
 
-	std::function<void(int, int)> OnHit; // called on collision
+	std::function<void(float, float)> OnHit; // called on collision
 };
